Antialiasing -> Color interpolation

+++ src/main/java/de/zintel/gfx/texture/ITexture.java
@@ -15,7 +15,7 @@ public interface ITexture {
 
 	int getHeight();
 
-	Color getColor(final int x, final int y);
+	Color getColor(final double x, final double y);
 
 	public static int mkTIdx(final double v, final int extent) {
 		return ((int) (Math.abs(v) * (extent - 1))) % extent;
